Print full pointer and fix issue with passing NULL pointers as argument

+++ main.h
@@ -30,7 +30,7 @@ int _print_digit(long int num);
 int _print_string(char *str);
 int _print_number(unsigned long n);
 int _print_hex_X(unsigned int num);
-int _print_pointer(unsigned int address);
+int _print_pointer(unsigned long int address);
 int _print_str_rev(char *str);
 int validityCheck(const char *format);
 char encodeChar(char c);
